pyproject.toml
[project]
name = "l1-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.25.0",
    "pillow>=11.3.0",
]
